backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "boto3>=1.40.64",
    "celery>=5.5.3",
    "email-validator>=2.2.0",
    "fastapi>=0.120.4",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "argon2-cffi>=23.1.0",
    "greenlet>=3.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-slugify>=8.0.4",
    "numpy>=1.26.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
    "psycopg2-binary>=2.9.9",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
